+++ docs/whatsnew.rst
@@ -9,7 +9,7 @@ The full release history can be viewed `at the GitHub yank releases page <https:
 0.22.1 Online Analysis Default
 ------------------------------
 - Online analysis will always run by default now, with no target error, run every checkpoint interval, and with at least 200 iterations
-- Online analysis can now be a set to the checkpoint interval by setting ``online_analysis_interval = "checkpoint"`` in the ``MultiStateSampler``
+- Online analysis can now be a set to the checkpoint interval by setting ``online_analysis_interval: "checkpoint"`` in the YAML files (application layer, not API)
 - Checkpoint interval increased from default of 10 to 200
 - Analysis now uses the online-analysis data if available by default